+++ api/v1/authentication.py
@@ -7,6 +7,8 @@ from app.users.schemas.user import UserLoginSchema
 from core.common.database import get_async_db_session
 from core.services.authention import AuthentionService
 from core.utils.bearer import get_access_token
+from core.services.view_information import ViewInformationService
+from core.utils.authorize import authorize_user
 
 authentication = APIRouter(prefix="/auth", tags=["Authentication"])
 
@@ -35,3 +37,16 @@ async def renew_token(
 ):
     auth_service = AuthentionService(db)
     return await auth_service.renew_token(refresh_access_token)
+
+
+@authentication.get("/me", status_code=status.HTTP_200_OK)
+async def view_me(
+    authorize=Depends(authorize_user("me:view")),
+    db_session: AsyncSession = Depends(get_async_db_session),
+):
+    view_service = ViewInformationService(db=db_session)
+    try:
+        return await view_service.get_user_detail(authorize.id)
+    except Exception as e:
+        print("Error in get_me:", e)
+        raise
